+++ grid2op/Environment/__init__.py
@@ -1,13 +1,13 @@
 __all__ = [
     "BaseEnv",
     "Environment",
-    "BaseMultiEnvironment",
-    "MultiEnvironment",
-    "MultiEnvironment2"
+    "BaseMultiProcessEnvironment",
+    "SingleEnvMultiProcess",
+    "MultiEnvMultiProcess"
 ]
 
 from grid2op.Environment.BaseEnv import BaseEnv
 from grid2op.Environment.Environment import Environment
-from grid2op.Environment.BaseMultiEnv import BaseMultiEnvironment
-from grid2op.Environment.MultiEnv import MultiEnvironment
-from grid2op.Environment.MultiEnv2 import MultiEnvironment2
+from grid2op.Environment.BaseMultiProcessEnv import BaseMultiProcessEnvironment
+from grid2op.Environment.SingleEnvMultiProcess import SingleEnvMultiProcess
+from grid2op.Environment.MultiEnvMultiProcess import MultiEnvMultiProcess