+++ grid2op/tests/test_MultiProcess.py
@@ -9,22 +9,22 @@
 import pdb
 import warnings
 from grid2op.tests.helper_path_test import *
-from grid2op.Environment import BaseMultiEnvironment
-from grid2op.Environment import MultiEnvironment
-from grid2op.Environment import MultiEnvironment2
+from grid2op.Environment import BaseMultiProcessEnvironment
+from grid2op.Environment import SingleEnvMultiProcess
+from grid2op.Environment import MultiEnvMultiProcess
 from grid2op.MakeEnv import make
 from grid2op.Observation import CompleteObservation
 import pdb
 
 
-class TestBaseMultiEnv(unittest.TestCase):
+class TestBaseMultiProcessEnvironment(unittest.TestCase):
     def test_creation_multienv(self):
         nb_env = 2
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
             with make("rte_case5_example", test=True) as env:
                 envs = [env for _ in range(nb_env)]
-                multi_envs = BaseMultiEnvironment(envs)
+                multi_envs = BaseMultiProcessEnvironment(envs)
 
         obss, rewards, dones, infos = multi_envs.step([env.action_space() for _ in range(multi_envs.nb_env)])
         for ob in obss:
@@ -48,26 +48,26 @@ class TestBaseMultiEnv(unittest.TestCase):
             with make("rte_case5_example", test=True) as env:
                 envs = [env for _ in range(nb_env)]
                 env.seed(2)
-                multi_envs1 = BaseMultiEnvironment(envs)
+                multi_envs1 = BaseMultiProcessEnvironment(envs)
                 seeds_1 = multi_envs1.get_seeds()
                 multi_envs1.close()
-                multi_envs2 = BaseMultiEnvironment(envs)
+                multi_envs2 = BaseMultiProcessEnvironment(envs)
                 seeds_2 = multi_envs2.get_seeds()
                 multi_envs2.close()
                 env.seed(2)
-                multi_envs3 = BaseMultiEnvironment(envs)
+                multi_envs3 = BaseMultiProcessEnvironment(envs)
                 seeds_3 = multi_envs3.get_seeds()
                 multi_envs3.close()
                 assert np.all(seeds_1 == seeds_3)
                 assert np.any(seeds_1 != seeds_2)
 
-class TestMultiEnv(unittest.TestCase):
+class TestSingleEnvMultiProcess(unittest.TestCase):
     def test_creation_multienv(self):
         nb_env = 2
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
             with make("rte_case5_example", test=True) as env:
-                multi_envs = MultiEnvironment(env=env, nb_env=nb_env)
+                multi_envs = SingleEnvMultiProcess(env=env, nb_env=nb_env)
 
         obss, rewards, dones, infos = multi_envs.step([env.action_space() for _ in range(multi_envs.nb_env)])
         for ob in obss:
@@ -90,27 +90,27 @@ class TestMultiEnv(unittest.TestCase):
             warnings.filterwarnings("ignore")
             with make("rte_case5_example", test=True) as env:
                 env.seed(2)
-                multi_envs1 = MultiEnvironment(env=env, nb_env=nb_env)
+                multi_envs1 = SingleEnvMultiProcess(env=env, nb_env=nb_env)
                 seeds_1 = multi_envs1.get_seeds()
                 multi_envs1.close()
-                multi_envs2 = MultiEnvironment(env=env, nb_env=nb_env)
+                multi_envs2 = SingleEnvMultiProcess(env=env, nb_env=nb_env)
                 seeds_2 = multi_envs2.get_seeds()
                 multi_envs2.close()
                 env.seed(2)
-                multi_envs3 = MultiEnvironment(env=env, nb_env=nb_env)
+                multi_envs3 = SingleEnvMultiProcess(env=env, nb_env=nb_env)
                 seeds_3 = multi_envs3.get_seeds()
                 multi_envs3.close()
                 assert np.all(seeds_1 == seeds_3)
                 assert np.any(seeds_1 != seeds_2)
 
-class TestMultiEnv2(unittest.TestCase):
+class TestMultiEnvMultiProcess(unittest.TestCase):
     def test_creation_multienv(self):
         nb_envs = [1, 1]
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
             with make("rte_case5_example", test=True) as env:
                 envs = [env for _ in range(len(nb_envs))]
-                multi_envs = MultiEnvironment2(envs, nb_envs)
+                multi_envs = MultiEnvMultiProcess(envs, nb_envs)
 
         obss, rewards, dones, infos = multi_envs.step([env.action_space() for _ in range(multi_envs.nb_env)])
         for ob in obss:
@@ -134,14 +134,14 @@ class TestMultiEnv2(unittest.TestCase):
             with make("rte_case5_example", test=True) as env:
                 envs = [env for _ in range(len(nb_envs))]
                 env.seed(2)
-                multi_envs1 = MultiEnvironment2(envs, nb_envs)
+                multi_envs1 = MultiEnvMultiProcess(envs, nb_envs)
                 seeds_1 = multi_envs1.get_seeds()
                 multi_envs1.close()
-                multi_envs2 = MultiEnvironment2(envs, nb_envs)
+                multi_envs2 = MultiEnvMultiProcess(envs, nb_envs)
                 seeds_2 = multi_envs2.get_seeds()
                 multi_envs2.close()
                 env.seed(2)
-                multi_envs3 = MultiEnvironment2(envs, nb_envs)
+                multi_envs3 = MultiEnvMultiProcess(envs, nb_envs)
                 seeds_3 = multi_envs3.get_seeds()
                 multi_envs3.close()
                 assert np.all(seeds_1 == seeds_3)